chore(sidebar): logic now in only one file
